+++ Source/Controller.py
@@ -23,6 +23,72 @@ from PDFreport import PDF
 
 
 class Controller:
+    
+
+    @staticmethod
+    def writeReport(title, fileName, pathOut, outFilePath, level):
+        print('Writing PDF Report...')
+        numLevelDict = {'L1A':1,'L1B':2,'L1C':3,'L1D':4,'L1E':5,'L2':6}
+        numLevel = numLevelDict[level]
+        
+        # Reports
+        reportPath = os.path.join(pathOut, 'Reports')
+        if os.path.isdir(reportPath) is False:
+            os.mkdir(reportPath)
+        fileName = fileName[0]
+        dirPath = os.getcwd()
+        inLogPath = os.path.join(dirPath, 'Logs')
+        inPlotPath = os.path.join(pathOut,'Plots')
+        # outPDF = os.path.join(reportPath,'Reports', f'{fileName}.pdf')
+        outPDF = os.path.join(reportPath, f'{fileName}_{level}.pdf')
+        # title = f'File: {fileName} Collected: {root.attributes["TIME-STAMP"]}'
+
+        pdf = PDF()
+        pdf.set_title(title)
+        pdf.set_author(f'HyperInSPACE_{MainConfig.settings["version"]}')
+
+        inLog = os.path.join(inLogPath,f'{fileName}_L1A.log')
+        if os.path.isfile(inLog):
+            # pdf.print_chapter(root,'L1A', 'Process RAW to L1A', inLog, inPlotPath, fileName, outFilePath)
+            pdf.print_chapter('L1A', 'Process RAW to L1A', inLog, inPlotPath, fileName, outFilePath)
+
+        if numLevel > 1:
+            inLog = os.path.join(inLogPath,f'{fileName}_L1A_L1B.log')
+            if os.path.isfile(inLog):
+                # pdf.print_chapter(root,'L1B', 'Process L1A to L1B', inLog, inPlotPath, fileName, outFilePath)
+                pdf.print_chapter('L1B', 'Process L1A to L1B', inLog, inPlotPath, fileName, outFilePath)
+        
+        if numLevel > 2:
+            inLog = os.path.join(inLogPath,f'{fileName}_L1B_L1C.log')
+            if os.path.isfile(inLog):
+                # pdf.print_chapter(root,'L1C', 'Process L1B to L1C', inLog, inPlotPath, fileName, outFilePath)
+                pdf.print_chapter('L1C', 'Process L1B to L1C', inLog, inPlotPath, fileName, outFilePath)
+        
+        if numLevel > 3:
+            inLog = os.path.join(inLogPath,f'{fileName}_L1C_L1D.log')
+            if os.path.isfile(inLog):
+                # pdf.print_chapter(root,'L1D', 'Process L1C to L1D', inLog, inPlotPath, fileName, outFilePath)
+                pdf.print_chapter('L1D', 'Process L1C to L1D', inLog, inPlotPath, fileName, outFilePath)
+        
+        if numLevel > 4:
+            inLog = os.path.join(inLogPath,f'{fileName}_L1D_L1E.log')
+            if os.path.isfile(inLog):
+                # pdf.print_chapter(root,'L1E', 'Process L1D to L1E', inLog, inPlotPath, fileName, outFilePath)
+                pdf.print_chapter('L1E', 'Process L1D to L1E', inLog, inPlotPath, fileName, outFilePath)
+        
+        if numLevel > 5:
+            inLog = os.path.join(inLogPath,f'{fileName}_L1E_L2.log')
+            if os.path.isfile(inLog):
+                # pdf.print_chapter(root,'L2', 'Process L1E to L2', inLog, inPlotPath, fileName, outFilePath)
+                pdf.print_chapter('L2', 'Process L1E to L2', inLog, inPlotPath, fileName, outFilePath)
+
+        try:
+            pdf.output(outPDF, 'F')
+        except:
+            msg = 'Unable to write the PDF file. It may be open in another program.'
+            Utilities.errorWindow("File Error", msg)
+            print(msg)
+            Utilities.writeLogFile(msg)
 
     @staticmethod
     def generateContext(calibrationMap):
@@ -116,23 +182,24 @@ class Controller:
         return ancillaryData
 
     @staticmethod
-    def processL1a(inFilePath, outFilePath, calibrationMap):      
+    def processL1a(inFilePath, outFilePath, calibrationMap): 
+        root = None     
         if not os.path.isfile(inFilePath):
             msg = 'No such file...'
             Utilities.errorWindow("File Error", msg)
             print(msg)
             Utilities.writeLogFile(msg)
-            return None
+            return None, 'Never'
 
         # Process the data
         msg = "ProcessL1a"
         print(msg)
-        # Utilities.writeLogFile(msg)
-        root = ProcessL1a.processL1a(inFilePath, outFilePath, calibrationMap)
+        root = ProcessL1a.processL1a(inFilePath, calibrationMap)
 
         # Apply SZA filter 
         if ConfigFile.settings["bL1aCleanSZA"]:
             if root is not None:
+                timeStamp = root.attributes['TIME-STAMP'] # a string
                 for gp in root.groups:                              
                     # try:
                     if 'FrameTag' in gp.attributes:
@@ -146,7 +213,7 @@ class Controller:
                                 msg = f'SZA too low. Discarding entire file. {round(90-np.nanmax(elevation))}'
                                 print(msg)
                                 Utilities.writeLogFile(msg)
-                                return None
+                                return None, timeStamp
                             else:
                                 msg = f'SZA passed filter: {round(90-np.nanmax(elevation))}'
                                 print(msg)
@@ -162,42 +229,47 @@ class Controller:
         # Write output file
         if root is not None:
             try:
-                root.writeHDF5(outFilePath)
+                root.writeHDF5(outFilePath)               
             except:                
                 msg = 'Unable to write L1A file. It may be open in another program.'
                 Utilities.errorWindow("File Error", msg)
                 print(msg)
                 Utilities.writeLogFile(msg)
-                return None
+                return None, timeStamp
         else:
             msg = "L1a processing failed. Nothing to output."
             if MainConfig.settings["popQuery"] == 0:
                 Utilities.errorWindow("File Error", msg)
             print(msg)
             Utilities.writeLogFile(msg)
-            return None
+            return None, 'Unknown'
+            
+        return root, timeStamp
 
     @staticmethod
-    def processL1b(inFilePath, outFilePath, calibrationMap):      
+    def processL1b(inFilePath, outFilePath, calibrationMap):    
+        root = None  
+        timeStamp = 'Unknown'
         if not os.path.isfile(inFilePath):
             print('No such input file: ' + inFilePath)
-            return None
+            return None, timeStamp
 
         # Process the data
         print("ProcessL1b")
         try:
             root = HDFRoot.readHDF5(inFilePath)
+            timeStamp = root.attributes['TIME-STAMP'] # a string
         except:
             msg = "Unable to open file. May be open in another application."
             Utilities.errorWindow("File Error", msg)
             print(msg)
             Utilities.writeLogFile(msg)
-            return None
+            return None, timeStamp
 
         root = ProcessL1b.processL1b(root, calibrationMap)
 
         # Write output file
-        if root is not None:
+        if root is not None:            
             try:
                 root.writeHDF5(outFilePath)
             except:
@@ -205,23 +277,26 @@ class Controller:
                 Utilities.errorWindow("File Error", msg)
                 print(msg)
                 Utilities.writeLogFile(msg)
-                return None
+                return None, timeStamp
         else:
             msg = "L1b processing failed. Nothing to output."
             if MainConfig.settings["popQuery"] == 0:
                 Utilities.errorWindow("File Error", msg)
             print(msg)
             Utilities.writeLogFile(msg)
-            return None
+            return None, timeStamp
+
+        return root, timeStamp
 
     @staticmethod
     def processL1c(inFilePath, outFilePath, ancillaryData=None):
-
+        root = None
+        timeStamp = 'Unknown'
         _,fileName = os.path.split(outFilePath)
         
         if not os.path.isfile(inFilePath):
             print('No such input file: ' + inFilePath)
-            return None
+            return None, timeStamp
 
         # Process the data
         msg = ("ProcessL1c: " + inFilePath)
@@ -229,17 +304,18 @@ class Controller:
         Utilities.writeLogFile(msg,'w')
         try:
             root = HDFRoot.readHDF5(inFilePath)
+            timeStamp = root.attributes['TIME-STAMP'] # a string
         except:
             msg = "Unable to open file. May be open in another application."
             Utilities.errorWindow("File Error", msg)
             print(msg)
             Utilities.writeLogFile(msg)
-            return None
+            return None, timeStamp
 
         root = ProcessL1c.processL1c(root, fileName, ancillaryData)     
 
         # Write output file
-        if root is not None:
+        if root is not None:            
             try:
                 root.writeHDF5(outFilePath)
             except:
@@ -247,21 +323,24 @@ class Controller:
                 Utilities.errorWindow("File Error", msg)
                 print(msg)
                 Utilities.writeLogFile(msg)
-                return None
+                return None, timeStamp
         else:
             msg = "L1c processing failed. Nothing to output."
             if MainConfig.settings["popQuery"] == 0:
                 Utilities.errorWindow("File Error", msg)
             print(msg)
             Utilities.writeLogFile(msg)
-            return None
+            return None, timeStamp
+        
+        return root, timeStamp
 
     @staticmethod
     def processL1d(inFilePath, outFilePath):        
-        
+        root = None
+        timeStamp = 'Unknown'
         if not os.path.isfile(inFilePath):
             print('No such input file: ' + inFilePath)
-            return None
+            return None, timeStamp
 
         # Process the data
         msg = ("ProcessL1d: " + inFilePath)
@@ -269,12 +348,13 @@ class Controller:
         Utilities.writeLogFile(msg,'w')
         try:
             root = HDFRoot.readHDF5(inFilePath)
+            timeStamp = root.attributes['TIME-STAMP'] # a string
         except:
             msg = "Unable to open file. May be open in another application."
             Utilities.errorWindow("File Error", msg)
             print(msg)
             Utilities.writeLogFile(msg)
-            return None
+            return None, timeStamp
 
         root = ProcessL1d.processL1d(root)     
 
@@ -282,38 +362,54 @@ class Controller:
         if root is not None:
             try:                
                 root.writeHDF5(outFilePath)
+
+                # Create Plots
+                # Radiometry
+                dirpath = os.getcwd()
+                _, filename = os.path.split(outFilePath)
+                if ConfigFile.settings['bL2PlotEs']==1:
+                    Utilities.plotRadiometryL1D(root, dirpath, filename, rType='ES')
+                if ConfigFile.settings['bL2PlotLi']==1:
+                    Utilities.plotRadiometryL1D(root, dirpath, filename, rType='LI')
+                if ConfigFile.settings['bL2PlotLt']==1:
+                    Utilities.plotRadiometryL1D(root, dirpath, filename, rType='LT')
             except:
                 msg = "Unable to write file. May be open in another application."
                 Utilities.errorWindow("File Error", msg)
                 print(msg)
                 Utilities.writeLogFile(msg)
-                return None
+                return None, timeStamp
         else:
             msg = "L1d processing failed. Nothing to output."
             if MainConfig.settings["popQuery"] == 0:
                 Utilities.errorWindow("File Error", msg)
             print(msg)
             Utilities.writeLogFile(msg)
-            return None            
+            return None, timeStamp    
+
+        return root, timeStamp
 
     @staticmethod
     def processL1e(inFilePath, outFilePath):
+        root = None
+        timeStamp = 'Unknown'
         _,fileName = os.path.split(outFilePath)
 
         if not os.path.isfile(inFilePath):
             print('No such input file: ' + inFilePath)
-            return None
+            return None, timeStamp
 
         # Process the data
         print("ProcessL1e")
         try:
             root = HDFRoot.readHDF5(inFilePath)
+            timeStamp = root.attributes['TIME-STAMP'] # a string
         except:
             msg = "Unable to open file. May be open in another application."
             Utilities.errorWindow("File Error", msg)
             print(msg)
             Utilities.writeLogFile(msg)
-            return None
+            return None, timeStamp
 
         root = ProcessL1e.processL1e(root, fileName)
 
@@ -326,22 +422,25 @@ class Controller:
                 Utilities.errorWindow("File Error", msg)
                 print(msg)
                 Utilities.writeLogFile(msg)
-                return None
+                return None, timeStamp
         else:
             msg = "L1e processing failed. Nothing to output."
             if MainConfig.settings["popQuery"] == 0:
                 Utilities.errorWindow("File Error", msg)
             print(msg)
             Utilities.writeLogFile(msg)
-            return None  
+            return None, timeStamp  
+        
+        return root, timeStamp
 
 
     @staticmethod
     def processL2(inFilePath, outFilePath, ancillaryData):
-        
+        root = None
+        timeStamp = 'Unknown'
         if not os.path.isfile(inFilePath):
             print('No such input file: ' + inFilePath)
-            return None, outFilePath
+            return None, outFilePath, timeStamp
 
         # Process the data
         msg = ("ProcessL2: " + inFilePath)
@@ -349,12 +448,13 @@ class Controller:
         Utilities.writeLogFile(msg,'w')
         try:
             root = HDFRoot.readHDF5(inFilePath)
+            timeStamp = root.attributes['TIME-STAMP'] # a string
         except:
             msg = "Unable to open file. May be open in another application."
             Utilities.errorWindow("File Error", msg)
             print(msg)
             Utilities.writeLogFile(msg)
-            return None, outFilePath
+            return None, outFilePath, timeStamp
         
         root.attributes['In_Filepath'] = inFilePath
         root = ProcessL2.processL2(root, ancillaryData)
@@ -396,20 +496,20 @@ class Controller:
         if root is not None:
             try:
                 root.writeHDF5(outFilePath)
-                return root, outFilePath
+                return root, outFilePath, timeStamp
             except:
                 msg = "Unable to write file. May be open in another application."
                 Utilities.errorWindow("File Error", msg)
                 print(msg)
                 Utilities.writeLogFile(msg)
-                return None, outFilePath
+                return None, outFilePath, timeStamp
         else:
             msg = "L2 processing failed. Nothing to output."
             if MainConfig.settings["popQuery"] == 0:
                 Utilities.errorWindow("File Error", msg)
             print(msg)
             Utilities.writeLogFile(msg)
-            return None, outFilePath
+            return None, outFilePath, timeStamp
 
     # Process every file in a list of files 1 level
     @staticmethod
@@ -430,65 +530,51 @@ class Controller:
             os.environ["LOGFILE"] = (fileName + '_' + level + '.log')
         msg = "Process Single Level"
         Utilities.writeLogFile(msg,mode='w')
+        
+        fileName = fileName.split('_') # [basefilename, level]
 
-        if level == "L1A" or \
-            level == "L1B" or \
-            level == "L1C" or \
-            level == "L1D":       
+        # Check for base output directory
+        if os.path.isdir(pathOut):
+            pathOutLevel = os.path.join(pathOut,level)
+        else:
+            msg = "Bad output destination. Select new Output Data Directory."
+            print(msg)
+            Utilities.writeLogFile(msg)
+            return False
 
-            if os.path.isdir(pathOut):
-                pathOut = os.path.join(pathOut,level)
-                if os.path.isdir(pathOut) is False:
-                    os.mkdir(pathOut)
-            else:
-                msg = "Bad output destination. Select new Output Data Directory."
-                print(msg)
-                Utilities.writeLogFile(msg)
-                return False
-            
-            if level == "L1A":               
-                outFilePath = os.path.join(pathOut,fileName + "_" + level + ".hdf")
-            else:
-                # split the level suffix off the infile to make the outfilename
-                fileName = fileName.split('_')
-                outFilePath = os.path.join(pathOut,fileName[0] + "_" + level + ".hdf")                
+        # Add output level directory if necessary
+        if os.path.isdir(pathOutLevel) is False:
+            os.mkdir(pathOutLevel)
+        
+        outFilePath = os.path.join(pathOutLevel,fileName[0] + "_" + level + ".hdf")        
+    
+        if level == "L1A" or level == "L1B" or \
+            level == "L1C" or level == "L1D":                           
             
             if level == "L1A":
-                Controller.processL1a(inFilePath, outFilePath, calibrationMap) 
+                root, timeStamp = Controller.processL1a(inFilePath, outFilePath, calibrationMap)                 
             elif level == "L1B":
-                Controller.processL1b(inFilePath, outFilePath, calibrationMap) 
+                root, timeStamp = Controller.processL1b(inFilePath, outFilePath, calibrationMap) 
             elif level == "L1C":
                 if ConfigFile.settings["bL1cSolarTracker"] == 0:
                     ancillaryData = Controller.processAncData(ancFile)
                 else:
                     ancillaryData = None
-                Controller.processL1c(inFilePath, outFilePath, ancillaryData)
+                root, timeStamp = Controller.processL1c(inFilePath, outFilePath, ancillaryData)
             elif level == "L1D":
-                Controller.processL1d(inFilePath, outFilePath) 
+                root, timeStamp = Controller.processL1d(inFilePath, outFilePath) 
 
+            # Confirm output file creation
             if os.path.isfile(outFilePath):
                 modTime = os.path.getmtime(outFilePath)
                 nowTime = datetime.datetime.now()
                 if nowTime.timestamp() - modTime < 60: # If the file exists and was created in the last minute...
                     msg = f'{level} file produced: \n {outFilePath}'
                     print(msg)
-                    Utilities.writeLogFile(msg)  
-                    return True                                        
+                    Utilities.writeLogFile(msg)                              
 
         elif level == "L1E":
-            if os.path.isdir(pathOut):
-                pathOut = os.path.join(pathOut,level)
-                if os.path.isdir(pathOut) is False:
-                    os.mkdir(pathOut)
-            else:
-                msg = "Bad output destination. Select new Output Data Directory."
-                print(msg)
-                Utilities.writeLogFile(msg)
-                return False
-
-            fileName = fileName.split('_')
-            outFilePath = os.path.join(pathOut,fileName[0] + "_" + level + ".hdf")
-            Controller.processL1e(inFilePath, outFilePath)   
+            root, timeStamp = Controller.processL1e(inFilePath, outFilePath)   
             
             if os.path.isfile(outFilePath):
                 modTime = os.path.getmtime(outFilePath)
@@ -506,22 +592,6 @@ class Controller:
                     return True
 
         elif level == "L2":          
-            if os.path.isdir(pathOut):
-                # Reports
-                reportPath = os.path.join(pathOut, 'Reports')
-                if os.path.isdir(reportPath) is False:
-                    os.mkdir(reportPath)
-
-                #Data
-                dataPath = os.path.join(pathOut,level)
-                if os.path.isdir(dataPath) is False:
-                    os.mkdir(dataPath)
-            else:
-                msg = "Bad output destination. Select new Output Data Directory."
-                print(msg)
-                Utilities.writeLogFile(msg)
-                return False
-
             if ConfigFile.settings["bL1cSolarTracker"]:
                 ancillaryData = Controller.processAncData(ancFile)
             else:
@@ -529,9 +599,7 @@ class Controller:
                 # and will be extracted from the ANCILLARY_NOTRACKER group later
                 ancillaryData = None
 
-            fileName = fileName.split('_')
-            outFilePath = os.path.join(dataPath,fileName[0] + "_" + level + ".hdf")
-            root, outFilePath = Controller.processL2(inFilePath, outFilePath, ancillaryData)  
+            root, outFilePath, timeStamp = Controller.processL2(inFilePath, outFilePath, ancillaryData)  
             
             if os.path.isfile(outFilePath):
                 # Ensure that the L2 on file is recent before continuing with 
@@ -549,51 +617,21 @@ class Controller:
                         print(msg)
                         Utilities.writeLogFile(msg)
                         SeaBASSWriter.outputTXT_Type2(outFilePath)
-                    # return True
-        
-                    # Write Report
-                    print('Writing PDF Report...')
-                    fileName = fileName[0]
-                    dirPath = os.getcwd()
-                    inLogPath = os.path.join(dirPath, 'Logs')
-                    inPlotPath = os.path.join(pathOut,'Plots')
-                    outPDF = os.path.join(pathOut,'Reports', f'{fileName}.pdf')
-                    title = f'File: {fileName} Collected: {root.attributes["TIME-STAMP"]}'
+                    # return True        
 
-                    pdf = PDF()
-                    pdf.set_title(title)
-                    pdf.set_author(f'HyperInSPACE_{MainConfig.settings["version"]}')
+        # In case of processing failure, write the report at this Process level
+        #   Use numeric level for writeReport
+        title = f'File: {fileName[0]} Collected: {timeStamp}'
+        if root is None:                        
+            Controller.writeReport(title,fileName, pathOut, outFilePath, level)
+            return False
 
-                    inLog = os.path.join(inLogPath,f'{fileName}_L1A.log')
-                    if os.path.isfile(inLog):
-                        pdf.print_chapter(root,'L1A', 'Process RAW to L1A', inLog, inPlotPath, fileName, outFilePath)
-                    inLog = os.path.join(inLogPath,f'{fileName}_L1A_L1B.log')
-                    if os.path.isfile(inLog):
-                        pdf.print_chapter(root,'L1B', 'Process L1A to L1B', inLog, inPlotPath, fileName, outFilePath)
-                    inLog = os.path.join(inLogPath,f'{fileName}_L1B_L1C.log')
-                    if os.path.isfile(inLog):
-                        pdf.print_chapter(root,'L1C', 'Process L1B to L1C', inLog, inPlotPath, fileName, outFilePath)
-                    inLog = os.path.join(inLogPath,f'{fileName}_L1C_L1D.log')
-                    if os.path.isfile(inLog):
-                        pdf.print_chapter(root,'L1D', 'Process L1C to L1D', inLog, inPlotPath, fileName, outFilePath)
-                    inLog = os.path.join(inLogPath,f'{fileName}_L1D_L1E.log')
-                    if os.path.isfile(inLog):
-                        pdf.print_chapter(root,'L1E', 'Process L1D to L1E', inLog, inPlotPath, fileName, outFilePath)
-                    inLog = os.path.join(inLogPath,f'{fileName}_L1E_L2.log')
-                    if os.path.isfile(inLog):
-                        pdf.print_chapter(root,'L2', 'Process L1E to L2', inLog, inPlotPath, fileName, outFilePath)
-
-                    try:
-                        pdf.output(outPDF, 'F')
-                    except:
-                        msg = 'Unable to write the PDF file. It may be open in another program.'
-                        Utilities.errorWindow("File Error", msg)
-                        print(msg)
-                        Utilities.writeLogFile(msg)
-
-        msg = f'Process Single Level: {outFilePath} - DONE'
+        msg = f'Process Single Level: {outFilePath} - SUCCESSFUL'
         print(msg)
         Utilities.writeLogFile(msg)
+        if level == "L2":
+            Controller.writeReport(title,fileName, pathOut, outFilePath, level)
+        return True  
 
 
     # Process every file in a list of files from L0 to L2